Final Commit For Now
I was half-wrong w/the solid angle approach and Epstein wrong. The integral doesn't blow up like log(1/r) but not because of solid angles, instead because the integrand seems to become more sharply peaked with smaller r. In fact this peak seems to be enough to kill the log(1/r) behavior.